added loginlogout methods in separate file
