+++ src/components/modals/Login.js
@@ -1,9 +1,10 @@
-import React, { useState } from "react";
+import { useState, useContext } from "react";
+import LoginLogout from "../../utils/LoginLogout";
 import SignUpModal from "./SignUp";
-import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../../firebase";
+import AuthContext from "../../utils/AuthContext"; // Import the AuthContext
 
 export default function LoginModal({ isOpen, closeModal }) {
+  const { isLoggedIn } = useContext(AuthContext);
   const [modalMode, setModalMode] = useState(false);
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
@@ -16,26 +17,10 @@ export default function LoginModal({ isOpen, closeModal }) {
     return null;
   }
 
-  const handleLoginForm = async () => {
-
-    if (!email || !password) {
-        alert("Please fill in all required fields");
-        return;
-    }
-
-    // Firebase authenticator to log in a user
-    await signInWithEmailAndPassword(auth, email, password)
-        .then((userCredential) => {
-            // Signed in 
-            const user = userCredential.user;
-            console.log(user, ": is now logged in!");
-
-        })
-        .catch((error) => {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-        });
-};
+  const loginInfo = {
+    email,
+    password,
+  }
 
   return (
     <>
@@ -68,12 +53,7 @@ export default function LoginModal({ isOpen, closeModal }) {
             value={password}
             onChange={(e) => setPassword(e.target.value)}
           />
-          <button
-            type="button"
-            onClick={handleLoginForm}
-            className="flex items-center justify-center h-12 px-6  w-64 bg-yellow-500 mt-8 rounded font-semibold text-sm text-black hover:bg-yellow-400">
-            Login
-          </button>
+          <LoginLogout email={loginInfo.email} password={loginInfo.password} />
           <div className="flex mt-6 justify-center text-xs">
             <a href="#" className="text-blue-499 hover:text-yellow-300">
               Forgot Password
